move node-list-tree to its own component
