+++ src/components/node-list-tree/node-list-tree-item/node-list-tree-item.js
@@ -3,8 +3,8 @@ import classnames from 'classnames';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
 import { TreeItem } from '@mui/x-tree-view';
-import Row from './components/row/row';
-import { getDataTestAttribute } from '../../utils/get-data-test-attribute';
+import Row from '../row/row';
+import { getDataTestAttribute } from '../../../utils/get-data-test-attribute';
 
 const arrowIconColor = '#8e8e90';
