+++ src/components/node-list-tree/row/row.scss
@@ -1,5 +1,5 @@
-@use '../../../../styles/variables' as var;
-@use '../../styles/variables';
+@use '../../../styles/variables' as var;
+@use '../../nodes-panel/styles/variables' as variables;
 
 .MuiTreeItem-iconContainer svg {
   z-index: var.$zindex-MuiTreeItem-icon;
